Exclude non-Rust submodules
This also has a list of modules which are in the rust-lang orgs but
which we don't want to include (notably, LLVM and associated projects).
This is primarily because these are fairly large, and orthogonal,
projects from the core of Rust. Though they are certainly important to
our success, we want to restrict our lists to just Rust contributors for
the time being.