Only consider submodules if they're in .gitmodules
A submodule listed in .gitmodules isn't actually necessarily present
in-tree; however, the reverse for our purposes is true: if we were to
walk the whole tree to find submodules, if they're not in the
.gitmodules then we can't do anything anyway (we don't have a URL).

As such, we just check all of the paths contained in .gitmodules. This
is significantly faster.