Fix attributing subtree bump commits to a single release
This changes the discovery strategy for by-release commits. Instead of trying to
guess the delta by looking at the commits landed on master from one commit to
the next, we instead find all commits reachable from the new tag and all commits
reachable from the previous tag, and delta that.

This avoids the (previously buggy) situation where a commit moves from a
submodule to in-tree (i.e., when moving things to subtrees) which caused us to
count it for that release. This is because the commit was newly reachable in the
mainline repository, even though it was pre-existing in the AuthorMap of the
previous version's submodules.

The new implementation is somewhat slower as we're doing significantly more
tree-walking, but ultimately still fairly fast. We don't care too much about our
performance right now.